changed star datatype

+++ backend/db/migrations/20221022041006-create-review.js
@@ -23,7 +23,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       stars: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT(2, 1)
       },
       createdAt: {
         allowNull: false,
